#22/ @Aline @Isa @Vanderlei @Rosa: Resolvendo responsividade de frase inicial
