+++ frontend/src/components/Banner/index.js
@@ -1,0 +1,16 @@
+import React from 'react'
+import Banner from '../images/banner.png'
+import BannerStyle from'./BannerStyles.css'
+
+
+function BannerComponent(){
+
+    return(       
+        <div className="div-banner">    
+            <img alt="5 Marias" className="img-banner" fixed="top"  src={Banner}></img>
+            <div className="titulo">
+                <h1 className="font">Não deixe o seu resíduo ir para o lugar errado. Faça o descarte correto.</h1>     
+            </div>  
+        </div>  
+    )
+}export default BannerComponent
