+++ frontend/src/components/Footer/FooterStyle.css
@@ -4,6 +4,7 @@
     color: whitesmoke;
     text-align: center;
     padding-top: 60px;  
+
 }
 
 .footerStyle{
