+++ frontend/src/components/NavBar/index.js
@@ -1,13 +1,13 @@
 import React from 'react'
 import {Navbar,Nav} from 'react-bootstrap'
 import NavigationStyles from "./NavBarStyle.css"
-import {Logotipo} from '../images/IconsComponents.js'
+import Logotipo from '../images/logo-5marias.png'
 
 
 function NavBarComponent(){
         return (
             <Navbar className= "NavBar" bg="light"  expand="lg"  sticky="top"  >
-               <Logotipo></Logotipo>
+                <img src={Logotipo} width="200px"></img>
                     <Navbar.Toggle aria-controls="basic-navbar-nav " />
                     <Navbar.Collapse id="basic-navbar-nav">
                     <Nav className="mr-auto mx-auto nb-cont">
