--- frontend/src/components/images/BannerStyles.css
@@ -1,23 +1,0 @@
-.banner-style{
- display: flex;
-}
-    
-.titulo{
-    display: flex;
-    position: absolute;
-    text-align: center;
-    padding-top: 15%;
-    padding-bottom: 10%;
-    right:20% ;
-    left: 20%;
-    font-size: xxx-large;
-    font-family: Arial, Helvetica, sans-serif;
-    color:#3F3D56;
-  
-}
-
-.img-banner{
-    display: flex;
-    width: 100%;
-   
-}
